Replaced test_n2_accumulate with test_n2_accumulate2
